* libmsip,minisip,libmutil,libmnetutil,libmikey: Compilation problem fixes   for gcc 3.4.
git-svn-id: svn://svn.minisip.org/minisip/trunk@877 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmnetutil/include/libmnetutil/NetworkFunctions.h
@@ -32,7 +32,7 @@ class NetworkFunctions{
 	public:
 		static vector<string> getAllInterfaces();
 		static string getInterfaceIPStr(string iface);
-		static string getHostHandlingService(string service, string domain, int32_t &ret_port);
+		static string getHostHandlingService(string service, string domain, uint16_t &ret_port);
 
 	private:
