+++ libmnetutil/source/NetworkFunctions.cxx
@@ -146,7 +146,7 @@ string NetworkFunctions::getInterfaceIPStr(string iface){
 }
 
 
-string NetworkFunctions::getHostHandlingService(string service, string domain, int32_t &ret_port){
+string NetworkFunctions::getHostHandlingService(string service, string domain, uint16_t &ret_port){
     string ret;
 #ifndef WIN32
 	int32_t port=-1;
